group-2/project-1/code/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
dashscope>=1.0
requests>=2.31
pytest>=8.0
pytest-cov>=5.0
